pyproject.toml
[tool.poetry]
authors = ["DOLMA <tech@dolma.ru>"]
description = "This is backend project for dolma"
name = "dolma_backend"
version = "0.0.0"

[tool.poetry.dependencies]
python = ">=3.10,<4"

# Poetry requirements, требуется для корректной сборки Docker контейнеров
certifi = "*"
requests = "*"
virtualenv = "*"

argon2-cffi = "*"  # https://github.com/hynek/argon2_cffi
celery = ">=5.2.3,<6"  # https://github.com/celery/celery
python-slugify = "^7"  # https://github.com/un33k/python-slugify
pytz = "*"  # https://github.com/stub42/pytz
redis = "^4.4"  # https://github.com/redis/redis-py

# Django
# ------------------------------------------------------------------------------
django = "^5"  # https://www.djangoproject.com/
# replaced by decouple due inactivity of authors
# django-environ = "^0.4.5"  # https://github.com/joke2k/django-environ
django-allauth = "^0.54"  # https://github.com/pennersr/django-allauth
django-celery-beat = "^2.4"  # https://github.com/celery/django-celery-beat
django-dbbackup = "^4.0.2"  # https://github.com/django-dbbackup/django-dbbackup/
django-extensions = "^3.2.1"  # https://github.com/django-extensions/django-extensions
django-filter = "^23.5"  # https://github.com/carltongibson/django-filter/tree/main
django-health-check = "^3.17"  # https://github.com/KristianOellegaard/django-health-check
django-macaddress = "^1.8"  # https://github.com/django-macaddress/django-macaddress
django-model-utils = "^4.3.1"  # https://github.com/jazzband/django-model-utils
django-phonenumber-field = "^7.0.1"  # https://github.com/stefanfoulis/django-phonenumber-field
django-redis = "^5.2"  # https://github.com/niwinz/django-redis
django-split-settings = "^1.2"  # https://github.com/sobolevn/django-split-settings
python-decouple = "^3.6"  # https://github.com/henriquebastos/python-decouple

# Security
# ------------------------------------------------------------------------------
# changes auto_generation key to bigint affects migrations
# django-axes = "^5.14"  # https://github.com/jazzband/django-axes
django-cors-headers = "^3.13"  # https://github.com/adamchainz/django-cors-headers/releases
django-csp = "^3.7"  # https://github.com/mozilla/django-csp
django-http-referrer-policy = "^1.1.1"  # https://github.com/ubernostrum/django-referrer-policy/

# Logging
# ------------------------------------------------------------------------------
structlog = "^22.3" # https://github.com/hynek/structlog

# Mailing
# ------------------------------------------------------------------------------
django-anymail = { extras = ["sendgrid"], version = "^9.0" }

# Django REST Framework
djangorestframework = "^3.15"  # https://github.com/encode/django-rest-framework
drf-extensions = "^0.7.1"  # https://github.com/chibisov/drf-extensions

gunicorn = "^20.1"  # https://pypi.org/project/gunicorn/
phonenumbers = "^8.13.3"  # https://github.com/daviddrysdale/python-phonenumbers
psycopg2-binary = "*" # https://github.com/psycopg/psycopg2
sentry-sdk = "^1.12.1"  # https://github.com/getsentry/sentry-python
pytest-factoryboy = "^2.5.1"  # https://pypi.org/project/pytest-factoryboy/
djangorestframework-simplejwt = "^5.2.2"  # https://github.com/jazzband/djangorestframework-simplejwt
rules = "^3.3"  # https://github.com/dfunckt/django-rules
reportlab = "4.2"  # https://pypi.org/project/reportlab/
requests-cache = "0.9.8"  # https://pypi.org/project/requests-cache/
django-taggit = "^4.0.0"  # https://github.com/jazzband/django-taggit/
xlrd = "^2.0.1"  # https://github.com/python-excel/xlrd/
django-stubs = "^1.13"  # https://github.com/typeddjango/django-stubs
drf-yasg = "^1.21.5"   # https://github.com/axnsan12/drf-yasg/
httpx = "^0.24.1"  # https://github.com/encode/httpx
docxtpl = "^0.16.7"  # https://pypi.org/project/docxtpl/
beautifulsoup4 = "^4.12.2"  # https://pypi.org/project/beautifulsoup4/
lxml = "^4.9.2"  # https://pypi.org/project/lxml/
openai = "^1.32.0"  # https://github.com/openai/openai-python
pylightxl = "^1.61"
bs4 = "^0.0.2"
llama-index = "^0.10.43"
llama-index-llms-ollama = "^0.1.5"
django-admin-charts = "^1.3.1"
django-admin-tools = "^0.9.3"
django-multiselectfield = {url = "https://github.com/goinnn/django-multiselectfield/archive/master.zip"}
gdrfile = "^0.0.1"
locust = "^2.29.0"

[tool.poetry.dev-dependencies]
django-coverage-plugin = "^3.0"  # https://github.com/nedbat/django_coverage_plugin
django-debug-toolbar = "^3.8.1"  # https://github.com/jazzband/django-debug-toolbar
django-extra-checks = "^0.13.3"  # https://github.com/kalekseev/django-extra-checks
django-migration-linter = "^4.1"  # https://github.com/3YOURMIND/django-migration-linter
django-querycount = "^0.7"  # https://pypi.org/project/django-querycount/
nplusone = "^1.0"  # https://github.com/jmcarp/nplusone

# Code quality
# ------------------------------------------------------------------------------
axblack = "*"  # https://github.com/axiros/axblack
flake8-django = "1.1"  # https://github.com/rocioar/flake8-django
flake8-logging-format = "^0.9"  # https://github.com/rocioar/flake8-django
flake8-pytest-style = "^1.6"  # https://github.com/m-burst/flake8-pytest-style
nitpick = "^0.32"  # https://github.com/andreoliwa/nitpick
wemake-python-styleguide = "^0.16"  # https://wemake-python-stylegui.de/en/latest/
flake8 = "^4.0.1"  # https://pypi.org/project/flake8/

# Testing
# ------------------------------------------------------------------------------
django-test-migrations = "^1.2"  # https://github.com/wemake-services/django-test-migrations
factory-boy = "^3.2.1"  # https://github.com/FactoryBoy/factory_boy
hypothesis = "^6.61"  # https://pypi.org/project/hypothesis/
pytest = "^7.2"  # https://github.com/pytest-dev/pytest
pytest-cov = "^4"  # https://github.com/pytest-dev/pytest-cov
pytest-deadfixtures = "^2.2.1"  # https://github.com/jllorencetti/pytest-deadfixtures
pytest-django = "^4.8.0"  # https://github.com/pytest-dev/pytest-django
pytest-randomly = "^3.12"  # https://github.com/pytest-dev/pytest-randomly
pytest-testmon = "^1.3"  # https://github.com/tarpas/pytest-testmon
pytest-timeout = "^2.1"  # https://github.com/pytest-dev/pytest-timeout/


# Type checking
# ------------------------------------------------------------------------------
mypy = "*"  # https://github.com/python/mypy
types-requests = "^2.28"  # https://pypi.org/project/types-requests/
# django-stubs-ext = "^0.2" # commented until we don't need/undstnd features provided

# Documentation
# ------------------------------------------------------------------------------
doc8 = "^1.1.1"  # https://github.com/pycqa/doc8
sphinx = "^5.3"  # https://github.com/sphinx-doc/sphinx
sphinx-autodoc-typehints = "^1.19.5"  # https://github.com/agronholm/sphinx-autodoc-typehints
tomlkit = "^0.11.6"  # https://github.com/sdispater/tomlkit

# Linting
# ------------------------------------------------------------------------------
dennis = "^1.1"  # https://github.com/willkg/dennis
dotenv-linter = "^0.4"  # https://github.com/wemake-services/dotenv-linter
dump-env = "^1.4"  # https://github.com/wemake-services/dump-env
ipython = "^8.7"  # https://pypi.org/project/ipython/
polint = "^0.4"  # https://github.com/ziima/polint
safety = "^2.3.4"  # https://github.com/pyupio/safety
yamllint = "^1.28"  # https://github.com/adrienverge/yamllint

# CI needs
# ------------------------------------------------------------------------------
importlib-metadata = "^4"  # https://github.com/python/importlib_metadata

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"
model-bakery = "^1.18.1"

[tool.black]
line-length = 80

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
